Added objectqualifier to providers/procedures.sql, fixed that COUNT(m.UserID) error in MSSQL2000, made fulltext.sql compatible with MSSQL2000.

+++ yafsrc/YetAnotherForum.NET/install/providers/procedures.sql
@@ -139,12 +139,12 @@ CREATE PROCEDURE [{databaseOwner}].[{objectQualifier}prov_createapplication]
 )
 AS
 BEGIN
-	SET @ApplicationID = (SELECT ApplicationID FROM {objectQualifier}prov_Application WHERE ApplicationNameLwd=LOWER(@ApplicationName))
+	SET @ApplicationID = (SELECT ApplicationID FROM [{databaseOwner}].[{objectQualifier}prov_Application] WHERE ApplicationNameLwd=LOWER(@ApplicationName))
 	
 	IF (@ApplicationID IS Null)
 	BEGIN
 		    SELECT  @ApplicationId = NEWID()
-            INSERT  {objectQualifier}prov_Application(ApplicationId, ApplicationName, ApplicationNameLwd)
+            INSERT  [{databaseOwner}].[{objectQualifier}prov_Application] (ApplicationId, ApplicationName, ApplicationNameLwd)
             VALUES  (@ApplicationId, @ApplicationName, LOWER(@Applicationname))
     END
 END 
@@ -165,7 +165,7 @@ BEGIN
 
 	EXEC [{databaseOwner}].[{objectQualifier}prov_CreateApplication] @ApplicationName, @ApplicationId OUTPUT
 	
-	UPDATE {objectQualifier}prov_Membership SET Password=@Password, PasswordSalt=@PasswordSalt,
+	UPDATE [{databaseOwner}].[{objectQualifier}prov_Membership] SET Password=@Password, PasswordSalt=@PasswordSalt,
 		PasswordFormat=@PasswordFormat, PasswordAnswer=@PasswordAnswer
 	WHERE UsernameLwd=LOWER(@Username) and ApplicationID=@ApplicationID;
 
@@ -185,7 +185,7 @@ BEGIN
 	
 	EXEC [{databaseOwner}].[{objectQualifier}prov_CreateApplication] @ApplicationName, @ApplicationId OUTPUT
 	
-	UPDATE {objectQualifier}prov_Membership SET PasswordQuestion=@PasswordQuestion, PasswordAnswer=@PasswordAnswer
+	UPDATE [{databaseOwner}].[{objectQualifier}prov_Membership] SET PasswordQuestion=@PasswordQuestion, PasswordAnswer=@PasswordAnswer
 	WHERE UsernameLwd=LOWER(@Username) and ApplicationID=@ApplicationID;
 
 END
@@ -212,7 +212,7 @@ BEGIN
 	IF @UserKey IS NULL
 		SET @UserKey = NEWID()
 		
-	INSERT INTO {objectQualifier}prov_Membership(UserID,ApplicationID,Username,UsernameLwd,Password,PasswordSalt,PasswordFormat,Email,EmailLwd,PasswordQuestion,PasswordAnswer,IsApproved)
+	INSERT INTO [{databaseOwner}].[{objectQualifier}prov_Membership] (UserID,ApplicationID,Username,UsernameLwd,Password,PasswordSalt,PasswordFormat,Email,EmailLwd,PasswordQuestion,PasswordAnswer,IsApproved)
 		VALUES (@UserKey, @ApplicationID,@Username, LOWER(@Username), @Password, @PasswordSalt, @PasswordFormat, @Email, LOWER(@Email), @PasswordQuestion, @PasswordAnswer, @IsApproved);
 END
 GO
@@ -267,9 +267,9 @@ BEGIN
     
 	CREATE TABLE #RowNumber (RowNumber int IDENTITY (1, 1),  UserID uniqueidentifier)
 	
-	INSERT INTO #RowNumber (UserID) SELECT m.UserID FROM {objectQualifier}prov_Membership m INNER JOIN {objectQualifier}prov_Application a ON m.ApplicationID = a.ApplicationID  WHERE a.ApplicationID = @ApplicationID AND m.EmailLwd = LOWER(@EmailAddress)
+	INSERT INTO #RowNumber (UserID) SELECT m.UserID FROM [{databaseOwner}].[{objectQualifier}prov_Membership] m INNER JOIN [{databaseOwner}].[{objectQualifier}prov_Application] a ON m.ApplicationID = a.ApplicationID  WHERE a.ApplicationID = @ApplicationID AND m.EmailLwd = LOWER(@EmailAddress)
 
-	SELECT m.*, r.RowNumber FROM {objectQualifier}prov_Membership m INNER JOIN #RowNumber r ON m.UserID = r.UserID WHERE r.RowNumber >= @PagingLowerBoundary AND r.RowNumber <= @PagingUpperBoundary;
+	SELECT m.*, r.RowNumber FROM [{databaseOwner}].[{objectQualifier}prov_Membership] m INNER JOIN #RowNumber r ON m.UserID = r.UserID WHERE r.RowNumber >= @PagingLowerBoundary AND r.RowNumber <= @PagingUpperBoundary;
     
 	SET @TotalRecords = (SELECT COUNT(RowNumber) FROM #RowNumber)
 	DROP TABLE #RowNumber
@@ -301,9 +301,9 @@ BEGIN
     
 	CREATE TABLE #RowNumber (RowNumber int IDENTITY (1, 1),  UserID uniqueidentifier)
 	
-	INSERT INTO #RowNumber (UserID) SELECT m.UserID FROM {objectQualifier}prov_Membership m INNER JOIN {objectQualifier}prov_Application a ON m.ApplicationID = a.ApplicationID WHERE a.ApplicationID = @ApplicationID AND m.UsernameLwd = LOWER(@Username)
+	INSERT INTO #RowNumber (UserID) SELECT m.UserID FROM [{databaseOwner}].[{objectQualifier}prov_Membership] m INNER JOIN [{databaseOwner}].[{objectQualifier}prov_Application] a ON m.ApplicationID = a.ApplicationID WHERE a.ApplicationID = @ApplicationID AND m.UsernameLwd = LOWER(@Username)
 
-	SELECT m.*, r.RowNumber FROM {objectQualifier}prov_Membership m INNER JOIN #RowNumber r ON m.UserID = r.UserID WHERE r.RowNumber >= @PagingLowerBoundary AND r.RowNumber <= @PagingUpperBoundary;
+	SELECT m.*, r.RowNumber FROM [{databaseOwner}].[{objectQualifier}prov_Membership] m INNER JOIN #RowNumber r ON m.UserID = r.UserID WHERE r.RowNumber >= @PagingLowerBoundary AND r.RowNumber <= @PagingUpperBoundary;
     
 	SET @TotalRecords = (SELECT COUNT(RowNumber) FROM #RowNumber)
 	DROP TABLE #RowNumber
@@ -334,9 +334,9 @@ BEGIN
     
 	CREATE TABLE #RowNumber (RowNumber int IDENTITY (1, 1),  UserID uniqueidentifier)
 	
-	INSERT INTO #RowNumber (UserID) SELECT m.UserID FROM {objectQualifier}prov_Membership m INNER JOIN {objectQualifier}prov_Application a ON m.ApplicationID = a.ApplicationID WHERE a.ApplicationID = @ApplicationID
+	INSERT INTO #RowNumber (UserID) SELECT m.UserID FROM [{databaseOwner}].[{objectQualifier}prov_Membership] m INNER JOIN [{databaseOwner}].[{objectQualifier}prov_Application] a ON m.ApplicationID = a.ApplicationID WHERE a.ApplicationID = @ApplicationID
 
-	SELECT m.*, r.RowNumber FROM {objectQualifier}prov_Membership m INNER JOIN #RowNumber r ON m.UserID = r.UserID WHERE r.RowNumber >= @PagingLowerBoundary AND r.RowNumber <= @PagingUpperBoundary;
+	SELECT m.*, r.RowNumber FROM [{databaseOwner}].[{objectQualifier}prov_Membership] m INNER JOIN #RowNumber r ON m.UserID = r.UserID WHERE r.RowNumber >= @PagingLowerBoundary AND r.RowNumber <= @PagingUpperBoundary;
     
 	SET @TotalRecords = (SELECT COUNT(RowNumber) FROM #RowNumber)
 	DROP TABLE #RowNumber
@@ -360,7 +360,7 @@ BEGIN
 	EXEC [{databaseOwner}].[{objectQualifier}prov_CreateApplication] @ApplicationName, @ApplicationID OUTPUT
 	
 	DECLARE @NumberActive int
-	SET @NumberActive = (SELECT COUNT(m.UserID) FROM {objectQualifier}prov_Membership m INNER JOIN {objectQualifier}prov_Application a ON m.ApplicationID = a.ApplicationID  WHERE a.ApplicationID = @ApplicationID AND m.LastLogin >= @ActivityDate)
+	SET @NumberActive = (SELECT COUNT(1) FROM [{databaseOwner}].[{objectQualifier}prov_Membership] m INNER JOIN [{databaseOwner}].[{objectQualifier}prov_Application] a ON m.ApplicationID = a.ApplicationID  WHERE a.ApplicationID = @ApplicationID AND m.LastLogin >= @ActivityDate)
     
     RETURN @NumberActive
 
@@ -381,14 +381,14 @@ BEGIN
 	EXEC [{databaseOwner}].[{objectQualifier}prov_CreateApplication] @ApplicationName, @ApplicationID OUTPUT
 	
 	IF (@UserKey IS NULL)
-		SELECT m.* FROM {objectQualifier}prov_Membership m WHERE m.UsernameLwd = LOWER(@Username) and m.ApplicationID = @ApplicationID
+		SELECT m.* FROM [{databaseOwner}].[{objectQualifier}prov_Membership] m WHERE m.UsernameLwd = LOWER(@Username) and m.ApplicationID = @ApplicationID
 	ELSE
-		SELECT m.* FROM {objectQualifier}prov_Membership m WHERE m.UserID = @UserKey and m.ApplicationID = @ApplicationID
+		SELECT m.* FROM [{databaseOwner}].[{objectQualifier}prov_Membership] m WHERE m.UserID = @UserKey and m.ApplicationID = @ApplicationID
 	
 	-- IF USER IS ONLINE DO AN UPDATE USER	
 	IF (@UserIsOnline = 1)
 	BEGIN
-		UPDATE {objectQualifier}prov_Membership SET LastActivity = GETDATE() WHERE UsernameLwd = LOWER(@Username) and ApplicationID = @ApplicationID
+		UPDATE [{databaseOwner}].[{objectQualifier}prov_Membership] SET LastActivity = GETDATE() WHERE UsernameLwd = LOWER(@Username) and ApplicationID = @ApplicationID
 	END		
 END
 GO
@@ -404,7 +404,7 @@ BEGIN
 	DECLARE @ApplicationID uniqueidentifier
 	EXEC [{databaseOwner}].[{objectQualifier}prov_CreateApplication] @ApplicationName, @ApplicationID OUTPUT
 
-	SELECT m.Username FROM {objectQualifier}prov_Membership m INNER JOIN {objectQualifier}prov_Application a ON m.ApplicationID = a.ApplicationID  WHERE a.ApplicationID = @ApplicationID AND m.EmailLwd = LOWER(@Email);
+	SELECT m.Username FROM [{databaseOwner}].[{objectQualifier}prov_Membership] m INNER JOIN [{databaseOwner}].[{objectQualifier}prov_Application] a ON m.ApplicationID = a.ApplicationID  WHERE a.ApplicationID = @ApplicationID AND m.EmailLwd = LOWER(@Email);
 END
 GO
 
@@ -425,7 +425,7 @@ BEGIN
 
 	EXEC [{databaseOwner}].[{objectQualifier}prov_CreateApplication] @ApplicationName, @ApplicationID OUTPUT
 	
-	UPDATE {objectQualifier}prov_Membership SET
+	UPDATE [{databaseOwner}].[{objectQualifier}prov_Membership] SET
 	Password = @Password,
 	PasswordSalt = @PasswordSalt,
 	PasswordFormat = @PasswordFormat,
@@ -448,7 +448,7 @@ BEGIN
 
 	EXEC [{databaseOwner}].[{objectQualifier}prov_CreateApplication] @ApplicationName, @ApplicationID OUTPUT
 	
-	UPDATE {objectQualifier}prov_Membership SET
+	UPDATE [{databaseOwner}].[{objectQualifier}prov_Membership] SET
 	IsLockedOut = 0,
 	FailedPasswordAttempts = 0
 	WHERE ApplicationID = @ApplicationID AND
@@ -482,11 +482,11 @@ BEGIN
 	-- Check for UniqueEmail
 	IF (@UniqueEmail = 1)
 	BEGIN
-		IF (EXISTS (SELECT 1 FROM {objectQualifier}prov_Membership m WHERE m.UserID != @UserKey AND m.EmailLwd=LOWER(@Email) AND m.ApplicationID=@ApplicationID) )
+		IF (EXISTS (SELECT 1 FROM [{databaseOwner}].[{objectQualifier}prov_Membership] m WHERE m.UserID != @UserKey AND m.EmailLwd=LOWER(@Email) AND m.ApplicationID=@ApplicationID) )
 			RETURN (2)
 	END
 	
-	UPDATE {objectQualifier}prov_Membership SET
+	UPDATE [{databaseOwner}].[{objectQualifier}prov_Membership] SET
 	Username = @Username,
 	UsernameLwd = LOWER(@Username),
 	Email = @Email,
@@ -520,11 +520,11 @@ BEGIN
 	
 	EXEC [{databaseOwner}].[{objectQualifier}prov_CreateApplication] @ApplicationName, @ApplicationID OUTPUT
 
-	SET @UserID = (SELECT UserID FROM {objectQualifier}prov_Membership m WHERE m.UsernameLwd=LOWER(@Username) AND m.ApplicationID = @ApplicationID)
-	SET @RoleID = (SELECT RoleID FROM {objectQualifier}prov_Role r WHERE r.RolenameLwd=LOWER(@Rolename) AND r.ApplicationID = @ApplicationID)
+	SET @UserID = (SELECT UserID FROM [{databaseOwner}].[{objectQualifier}prov_Membership] m WHERE m.UsernameLwd=LOWER(@Username) AND m.ApplicationID = @ApplicationID)
+	SET @RoleID = (SELECT RoleID FROM [{databaseOwner}].[{objectQualifier}prov_Role] r WHERE r.RolenameLwd=LOWER(@Rolename) AND r.ApplicationID = @ApplicationID)
 	
-	IF (NOT EXISTS(SELECT 1 FROM {objectQualifier}prov_RoleMembership rm WHERE rm.UserID=@UserID AND rm.RoleID=@RoleID))
-		INSERT INTO {objectQualifier}prov_RoleMembership(RoleID, UserID) VALUES (@RoleID, @UserID);
+	IF (NOT EXISTS(SELECT 1 FROM [{databaseOwner}].[{objectQualifier}prov_RoleMembership] rm WHERE rm.UserID=@UserID AND rm.RoleID=@RoleID))
+		INSERT INTO [{databaseOwner}].[{objectQualifier}prov_RoleMembership] (RoleID, UserID) VALUES (@RoleID, @UserID);
 END
 GO
 
@@ -543,18 +543,18 @@ BEGIN
 	EXEC [{databaseOwner}].[{objectQualifier}prov_CreateApplication] @ApplicationName, @ApplicationID OUTPUT	
 	
 	SET @ErrorCode = 0
-	SET @RoleID = (SELECT RoleID FROM {objectQualifier}prov_Role r WHERE r.RolenameLwd=LOWER(@Rolename) AND r.ApplicationID = @ApplicationID)
+	SET @RoleID = (SELECT RoleID FROM [{databaseOwner}].[{objectQualifier}prov_Role] r WHERE r.RolenameLwd=LOWER(@Rolename) AND r.ApplicationID = @ApplicationID)
 	
 	IF (@DeleteOnlyIfRoleIsEmpty <> 0)
 	BEGIN
-		IF (EXISTS (SELECT 1 FROM {objectQualifier}prov_RoleMembership rm WHERE rm.RoleID=@RoleID))
+		IF (EXISTS (SELECT 1 FROM [{databaseOwner}].[{objectQualifier}prov_RoleMembership] rm WHERE rm.RoleID=@RoleID))
 			SELECT @ErrorCode = 2
 	ELSE
-		DELETE FROM {objectQualifier}prov_RoleMembership WHERE RoleID=@RoleID
+		DELETE FROM [{databaseOwner}].[{objectQualifier}prov_RoleMembership] WHERE RoleID=@RoleID
 	END	
 
 	IF (@ErrorCode = 0)
-		DELETE FROM {objectQualifier}prov_Role WHERE RoleID=@RoleID
+		DELETE FROM [{databaseOwner}].[{objectQualifier}prov_Role] WHERE RoleID=@RoleID
     
     RETURN @ErrorCode	
 END 
@@ -573,9 +573,9 @@ BEGIN
 	
 	EXEC [{databaseOwner}].[{objectQualifier}prov_CreateApplication] @ApplicationName, @ApplicationID OUTPUT
 
-	SET @RoleID = (SELECT RoleID FROM {objectQualifier}prov_Role r INNER JOIN {objectQualifier}prov_Application a ON r.ApplicationID = a.ApplicationID WHERE r.RolenameLwd=LOWER(@Rolename) AND a.ApplicationID = @ApplicationID)
+	SET @RoleID = (SELECT RoleID FROM [{databaseOwner}].[{objectQualifier}prov_Role] r INNER JOIN [{databaseOwner}].[{objectQualifier}prov_Application] a ON r.ApplicationID = a.ApplicationID WHERE r.RolenameLwd=LOWER(@Rolename) AND a.ApplicationID = @ApplicationID)
 
-	SELECT m.* FROM {objectQualifier}prov_Membership m INNER JOIN {objectQualifier}prov_RoleMembership rm ON m.UserID = rm.UserID WHERE rm.RoleID = @RoleID
+	SELECT m.* FROM [{databaseOwner}].[{objectQualifier}prov_Membership] m INNER JOIN [{databaseOwner}].[{objectQualifier}prov_RoleMembership] rm ON m.UserID = rm.UserID WHERE rm.RoleID = @RoleID
 		
 END
 GO
@@ -591,8 +591,8 @@ BEGIN
 	
 	EXEC [{databaseOwner}].[{objectQualifier}prov_CreateApplication] @ApplicationName, @ApplicationID OUTPUT
 	
-	IF (NOT EXISTS(SELECT 1 FROM {objectQualifier}prov_Role r WHERE r.ApplicationID = @ApplicationID AND r.RolenameLwd = LOWER(@Rolename)))
-		INSERT INTO {objectQualifier}prov_Role(RoleID, ApplicationID, RoleName, RoleNameLwd) VALUES (NEWID(),@ApplicationID, @Rolename,LOWER(@Rolename));		
+	IF (NOT EXISTS(SELECT 1 FROM [{databaseOwner}].[{objectQualifier}prov_Role] r WHERE r.ApplicationID = @ApplicationID AND r.RolenameLwd = LOWER(@Rolename)))
+		INSERT INTO [{databaseOwner}].[{objectQualifier}prov_Role] (RoleID, ApplicationID, RoleName, RoleNameLwd) VALUES (NEWID(),@ApplicationID, @Rolename,LOWER(@Rolename));		
 END
 GO
 
@@ -608,7 +608,7 @@ BEGIN
  	EXEC [{databaseOwner}].[{objectQualifier}prov_CreateApplication] @ApplicationName, @ApplicationID OUTPUT
 
 	IF (@Username is null)
-		SELECT r.* FROM {objectQualifier}prov_Role r WHERE r.ApplicationID = @ApplicationID
+		SELECT r.* FROM [{databaseOwner}].[{objectQualifier}prov_Role] r WHERE r.ApplicationID = @ApplicationID
 	ELSE
 		SELECT
 			r.*
@@ -636,9 +636,9 @@ BEGIN
 
 	EXEC [{databaseOwner}].[{objectQualifier}prov_CreateApplication] @ApplicationName, @ApplicationID OUTPUT
 
-	SELECT m.* FROM {objectQualifier}prov_RoleMembership rm 
-		INNER JOIN {objectQualifier}prov_Membership m ON rm.UserID = m.UserID
-		INNER JOIN {objectQualifier}prov_Role r ON rm.RoleID = r.RoleID
+	SELECT m.* FROM [{databaseOwner}].[{objectQualifier}prov_RoleMembership] rm 
+		INNER JOIN [{databaseOwner}].[{objectQualifier}prov_Membership] m ON rm.UserID = m.UserID
+		INNER JOIN [{databaseOwner}].[{objectQualifier}prov_Role] r ON rm.RoleID = r.RoleID
 		WHERE m.UsernameLwd=LOWER(@Username) AND r.RolenameLwd =LOWER(@Rolename) AND r.ApplicationID = @ApplicationID;
 END 
 GO
@@ -657,10 +657,10 @@ BEGIN
 
 	EXEC [{databaseOwner}].[{objectQualifier}prov_CreateApplication] @ApplicationName, @ApplicationID OUTPUT	
 	
-	SET @RoleID = (SELECT RoleID FROM {objectQualifier}prov_Role r WHERE r.RolenameLwd = LOWER(@Rolename) AND r.ApplicationID = @ApplicationID)
-	SET @UserID = (SELECT UserID FROM {objectQualifier}prov_Membership m WHERE m.UsernameLwd=LOWER(@Username) AND m.ApplicationID = @ApplicationID)
+	SET @RoleID = (SELECT RoleID FROM [{databaseOwner}].[{objectQualifier}prov_Role] r WHERE r.RolenameLwd = LOWER(@Rolename) AND r.ApplicationID = @ApplicationID)
+	SET @UserID = (SELECT UserID FROM [{databaseOwner}].[{objectQualifier}prov_Membership] m WHERE m.UsernameLwd=LOWER(@Username) AND m.ApplicationID = @ApplicationID)
 	
-	DELETE FROM {objectQualifier}prov_RoleMembership WHERE RoleID = @RoleID AND UserID=@UserID
+	DELETE FROM [{databaseOwner}].[{objectQualifier}prov_RoleMembership] WHERE RoleID = @RoleID AND UserID=@UserID
 	
 END 
 GO
@@ -676,7 +676,7 @@ BEGIN
 
 	EXEC [{databaseOwner}].[{objectQualifier}prov_CreateApplication] @ApplicationName, @ApplicationID OUTPUT
 	
-	SELECT COUNT(1) FROM {objectQualifier}prov_Role
+	SELECT COUNT(1) FROM [{databaseOwner}].[{objectQualifier}prov_Role]
 		WHERE RolenameLwd = LOWER(@Rolename);
 END 
 GO
